backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
channels>=4.0.0
channels-redis>=4.1.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0
daphne>=4.0.0
